Refactor search arguments
To address issue #1, I've refactored the CLI to take a more generic set
of arguments: a url, an extension uuid, or keyword(s).

If it's a url, we'll parse it to get the uuid. Otherwise, we pass the
arguments to our search and get the first result (ranked by popularity).

Also added `--dry-run` so you can use it as a search tool to see what
*would* be installed, which is useful if you're entering keywords.